Fix unsigned underflow in SDR perf time difference     macro
   SDR_PERF_TIME_DIFF_US used unsigned arithmetic (1000000UL), causing
   negative time differences from non-monotonic gettimeofday (e.g. NTP
   adjustments) to wrap to near-UINT64_MAX values. This produced bogus
   max xn time readings like 18446744073709550993 (which is 2^64 - 623).

   Use signed long arithmetic in the macro and clamp negative results to
   zero in sdr_perf_xn_end, sdr_perf_read_end, and sdr_perf_write_end.

+++ ici/include/sdr_perf.h
@@ -62,11 +62,13 @@ typedef struct
 	SdrPerfCaller	callers[SDR_PERF_CALLER_BUCKETS];
 } SdrPerfCounters;
 
-/*	Helper macro for time difference in microseconds.		*/
+/*	Helper macro for time difference in microseconds (signed).
+ *	Result is a long; callers must clamp negative values to 0
+ *	since gettimeofday is not monotonic and may go backwards.	*/
 
 #define SDR_PERF_TIME_DIFF_US(start, end) \
-	(((end).tv_sec - (start).tv_sec) * 1000000UL + \
-	 ((end).tv_usec - (start).tv_usec))
+	(((long)((end).tv_sec - (start).tv_sec)) * 1000000L + \
+	 ((long)(end).tv_usec - (long)(start).tv_usec))
 
 /*	Function prototypes.						*/
